Properly pack @markprompt/react (#47)
The `markprompt` package was renamed to `@markprompt/react`. It is now
ESM-only, and TypeScript is configured accordingly.

The package directory was renamed from `markprompt-react` to `react`.
This allows us to map paths using a glob when adding more packages.

The package is now built using `tsc --build` instead of a custom build
script.